requirements.txt
streamlit>=1.30.0
gensim>=4.3.0
plotly>=5.18.0
umap-learn>=0.5.5
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.0.0
